Enable vtable verification feature
This allows the -fvtable-verify flag to be used in the resulting
C++ compiler, which may be useful in debugging C++ crashes.
If GCC is built without the --enable-vtable-verify configure flag
then attempts to use -fvtable-verify are met with errors like
"-fvtable-verify=std is not supported in this configuration".

I also documented why other configure options are there.