Hardcode a version number for release
I've wasted a while looking for a better way to do this, but Omnibus
doesn't seem to provide a suitable mechanism.

It's possible to set build_version from the Git tag of this repo, but
we can't set build_iteration that way.

It's also possible to set build_version from the version of the GCC
component, but that requires that we're building a ref whose name
corresponds to the version number which isn't possible for us.

Oh well.

For nightly and developer builds we can add a mechanism that allows
overriding the package version by setting an environment variable
for the `omnibus build` command.